Use container functional components for easier context provision
